separated autocomplete functions into separate file and adjusted commands accordingly
