+++ main.py
@@ -8,6 +8,7 @@ import config
 import statistics
 import spotify_integration as spotify
 import homework
+import autocomplete as ac
 
 
 # set up Discord Bot with read and write message privileges, but without mentioning privileges
@@ -536,123 +537,6 @@ def get_top_albums_formatted(top_number: int = 5, min_ratings: int = 2, year: in
     return final_string
 
 
-# takes in year, minratings, and numalbums and returns the missing one (either minratings or numalbums)
-def top_albums_autocomplete_helper(whoscallin: str, year=datetime.now().year, minratings: int = None, numalbums: int = None):
-    ratings_year = {key: value for key, value in get_all_ratings().items() if year == -1 or year == key[2]}
-    if "maxalbums" in whoscallin:
-        if minratings is None:
-            return len(ratings_year)
-        final_dict = {key: value for key, value in ratings_year.items() if len(value) >= minratings}
-        return len(final_dict)
-    if "minrating" in whoscallin:
-        # sort the dictionary by number of ratings,
-        # then go down however many rows and return the number of ratings on that album
-        final_dict = dict(sorted(ratings_year.items(), key=lambda row: len(row[1]), reverse=True))
-        if numalbums is None:
-            return len(list(final_dict.values())[len(final_dict)-1])
-        return len(list(final_dict.values())[numalbums-1])
-    return
-
-
-# AUTOCOMPLETE AND CHOICES SECTION-----------------------------------------------------------------------------------
-# autocomplete names fail if a choice is over 100 characters, so this will modify names to take that into account
-def autocomplete_slice_names_100(name):
-    length = len(name)
-    if length <= 100:
-        return name
-    # here's how we solve this problem. find the rightmost space character. slice there.
-    # find the length of the right string. take the substring of the first string
-    # so that the length of that plus and length of the right string add to 97
-    r_index = 85
-    cut_title = name[:r_index + 1] + "..." + name[-1 * (97 - r_index) + 1:]
-    print(cut_title)
-    return cut_title
-
-
-# some of these functions wont work properly if there aren't exact matches which is bad, so this is a search function
-# it goes through each word in a string and tries to find if there are matching words in the current string
-def autocomplete_search_list(current_str: str, strs_to_search: tuple):
-    translate_table = str.maketrans('', '', '\'",-.!/():')
-    current_words = current_str.translate(translate_table).split(' ')
-    return_list = list()
-    for item in strs_to_search:
-        tuple_to_search = tuple(item[0].translate(translate_table).split(' '))
-        if all([any([string1.lower() in string2.lower() for string2 in tuple_to_search]) for string1 in current_words]):
-            return_list.append(item)
-    return return_list
-
-
-# this takes in a list of (artist, album, value) and returns that list with all entries as less than 100 characters
-def autocomplete_slice_list_names(choices):
-    return [tuple(autocomplete_slice_names_100(name) if len(name) > 100 else name for name in choice) for choice in choices]
-
-
-# gets a list of artists in album_master and returns a list of choices for use in autocomplete
-async def autocomplete_artist(interaction: discord.Interaction, current: str) -> list[app_commands.Choice[str]]:
-    final_list = {", ".join(row[0]) for row in get_album_master()}
-    return [Choice(name=artist, value=artist)
-            for artist in final_list if strip_names(current)[0] in strip_names(artist)[0]][:25]
-
-
-# same as above, but it does albums with autocomplete
-async def autocomplete_album(interaction: discord.Interaction, current: str) -> list[app_commands.Choice[str]]:
-    albums = tuple((row[1],) for row in get_album_master())
-    final_list = autocomplete_slice_list_names(albums)
-    return [Choice(name=album[0], value=album[0])
-            for album in final_list if strip_names(current)[0] in strip_names(album[0])[0]][:25]
-
-
-# gets a formatted list of choices of artist - album using spotify search
-async def autocomplete_spotify(interaction: discord.Interaction, current: str) -> list[app_commands.Choice[str]]:
-    results = await spotify.search_album(current)
-    choices = [(", ".join(tuple(artist.name for artist in entry.artists)) + f" ({entry.type}) - {entry.name}", f"{entry.id}") for entry in results]
-    final_choices = autocomplete_slice_list_names(choices)
-    return [Choice(name=f"{choice[0]}", value=f"{choice[1]}") for choice in final_choices][:25]
-
-
-# gets a formatted list of choices of artist - album using album_master
-async def autocomplete_artist_album(interaction: discord.Interaction, current: str) -> list[app_commands.Choice[str]]:
-    user_albums = tuple([(", ".join(row[0]) + f" - {row[1]}", f"{row[2]}") for row in get_album_master()])
-    half_final_list = autocomplete_search_list(current, user_albums)
-    final_list = autocomplete_slice_list_names(half_final_list)
-    return [Choice(name=entry[0], value=entry[1]) for entry in final_list][:25]
-
-
-# gets formatted choices for artist/album when editing/deleting rows from the list
-async def autocomplete_artist_album_user_specific(interaction: discord.Interaction, current: str) -> list[app_commands.Choice[str]]:
-    albums = [(", ".join(row[0]) + f" - {row[1]}", f"{row[3]}") for row in get_rows_from_user(interaction.user.id)]
-    final_list = autocomplete_slice_list_names(albums)
-    return [Choice(name=entry[0], value=entry[1])
-            for entry in final_list if current.lower() in entry[0].lower()][:25]
-
-
-# gets formatted choices for artist/album when editing/deleting rows from homework
-async def autocomplete_artist_album_homework_specific(interaction: discord.Interaction, current: str) -> list[app_commands.Choice[str]]:
-    homework_list = [(", ".join(row[3]) + f" - {row[4]}", f"{row[5]}")
-                     for row in homework.get_homework(conn=conn, user_id=interaction.user.id)]
-    final_list = autocomplete_slice_list_names(homework_list)
-    return [Choice(name=entry[0], value=entry[1])
-            for entry in final_list if current.lower() in entry[0].lower()][:25]
-
-
-# gets choices for num albums as part of top_albums command
-async def autocomplete_top_albums_numalbums(interaction: discord.Interaction, current: str) -> list[app_commands.Choice[int]]:
-    min_ratings = interaction.namespace.minimumratings
-    year = interaction.namespace.year if interaction.namespace.year is not None else datetime.now().year
-    max_num_albums = top_albums_autocomplete_helper("maxalbums", year=year, minratings=min_ratings)
-    return [Choice(name=str(num), value=num)
-            for num in range(1, max_num_albums+1) if str(current) in str(num)]
-
-
-# gets choices for num albums as part of top_albums command
-async def autocomplete_top_albums_minratings(interaction: discord.Interaction, current: str) -> list[app_commands.Choice[int]]:
-    numalbums = interaction.namespace.numberofalbums
-    year = interaction.namespace.year if interaction.namespace.year is not None else datetime.now().year
-    max_min_num_ratings = top_albums_autocomplete_helper("minrating", year=year, numalbums=numalbums)
-    return [Choice(name=str(num), value=num)
-            for num in range(1, max_min_num_ratings+1) if str(current) in str(num)]
-
-
 # whenever the bot is ready, it'll log this
 @client.event
 async def on_ready():
@@ -676,7 +560,11 @@ async def update(interaction: discord.Interaction):
 @app_commands.describe(year="the year of which you want the ratings")
 async def get_ratings(interaction: discord.Interaction, year: int = datetime.now().year):
     try:
-        await interaction.response.send_message(await get_rankings_message(year=year))
+        await interaction.response.defer()
+        message = await get_rankings_message(year=year)
+        messages = split_message(message)
+        await interaction.followup.send(messages[0])
+        [await interaction.channel.send(content=i) for i in messages[1:]]
     except Exception as error:
         print_exc()
         await interaction.response.send_message(content=error)
@@ -687,7 +575,7 @@ async def get_ratings(interaction: discord.Interaction, year: int = datetime.now
 @app_commands.describe(artist="the name of the artist",
                        album="the name of the album",
                        rating="the rating of the album")
-@app_commands.autocomplete(artist=autocomplete_artist, album=autocomplete_album)
+@app_commands.autocomplete(artist=ac.autocomplete_artist(get_album_master), album=ac.autocomplete_album(get_album_master))
 async def addmanual(interaction: discord.Interaction, artist: str, album: str, rating: float):
     try:
         artist = tuple(artist.split(","))
@@ -702,7 +590,7 @@ async def addmanual(interaction: discord.Interaction, artist: str, album: str, r
 # ADD COMMAND - uses spotify + autocomplete to find albums
 @tree.command(name='add_rating', description='add an album to your rankings with the help of spotify search', guild=my_guild)
 @app_commands.describe(searchkeywords="type in keywords for your search here")
-@app_commands.autocomplete(searchkeywords=autocomplete_spotify)
+@app_commands.autocomplete(searchkeywords=ac.autocomplete_spotify(spotify.search_album))
 async def add(interaction: discord.Interaction, searchkeywords: str, rating: float):
     try:
         artist, album, album_id, date, album_cover = spotify.get_album(album_id=searchkeywords)
@@ -740,7 +628,7 @@ async def addbulk(interaction: discord.Interaction, albums: str):
 @tree.command(name='edit_rating', description='edit a rating on an album', guild=my_guild)
 @app_commands.describe(entry="the artist - album whos rating you want to change",
                        rating="your new rating of the album (0-10)")
-@app_commands.autocomplete(entry=autocomplete_artist_album_user_specific)
+@app_commands.autocomplete(entry=ac.autocomplete_artist_album_user_specific(get_rows_from_user))
 async def edit(interaction: discord.Interaction, entry: str, rating: float):
     try:
         artist, album, album_id, date, image = spotify.get_album(album_id=entry)
@@ -759,7 +647,7 @@ async def edit(interaction: discord.Interaction, entry: str, rating: float):
 # REMOVE COMMAND - removes an album from a users list
 @tree.command(name="remove_rating", description="remove an album from your ranking", guild=my_guild)
 @app_commands.describe(entry="the artist - album you want to remove (see autocomplete)")
-@app_commands.autocomplete(entry=autocomplete_artist_album_user_specific)
+@app_commands.autocomplete(entry=ac.autocomplete_artist_album_user_specific(get_rows_from_user))
 async def remove(interaction: discord.Interaction, entry: str):
     try:
         artist, album, album_id, date, image = spotify.get_album(album_id=entry)
@@ -776,7 +664,7 @@ async def remove(interaction: discord.Interaction, entry: str):
 # COVER COMMAND - displays the cover of the album
 @tree.command(name="album_cover", description="displays the cover of an album", guild=my_guild)
 @app_commands.describe(entry="the album - artist you want to see the cover of (see autocomplete)")
-@app_commands.autocomplete(entry=autocomplete_spotify)
+@app_commands.autocomplete(entry=ac.autocomplete_spotify(spotify.search_album))
 async def cover(interaction: discord.Interaction, entry: str):
     try:
         artist, album, album_id, date, album_cover = spotify.get_album(album_id=entry)
@@ -791,7 +679,7 @@ async def cover(interaction: discord.Interaction, entry: str):
 # STATS COMMAND - displays stats for a certain album based on current rankings
 @tree.command(name='stats', description='find out stats about an album', guild=my_guild)
 @app_commands.describe(entry="the artist - album you are trying to get (see autocomplete)")
-@app_commands.autocomplete(entry=autocomplete_artist_album)
+@app_commands.autocomplete(entry=ac.autocomplete_artist_album(get_album_master))
 async def stats(interaction: discord.Interaction, entry: str):
     try:
         artist, album, album_id, date, image = spotify.get_album(album_id=entry)
@@ -810,7 +698,8 @@ async def stats(interaction: discord.Interaction, entry: str):
 @app_commands.describe(numberofalbums="how many albums do you want to see ranked? (default: 5)",
                        minimumratings="how many rankings do you want the album to have minimum (default: 1)",
                        year="the year you want to filter by, -1 for no filtering")
-@app_commands.autocomplete(numberofalbums=autocomplete_top_albums_numalbums, minimumratings=autocomplete_top_albums_minratings)
+@app_commands.autocomplete(numberofalbums=ac.autocomplete_top_albums_numalbums(get_all_ratings),
+                           minimumratings=ac.autocomplete_top_albums_minratings(get_all_ratings))
 @app_commands.choices(sortby=[Choice(name='average', value='avg'), Choice(name='standard deviation', value='std')])
 async def top_albums(interaction: discord.Interaction, numberofalbums: int = 5, minimumratings: int = 1, sortby: str = 'avg', year: int = datetime.now().year):
     try:
@@ -823,8 +712,8 @@ async def top_albums(interaction: discord.Interaction, numberofalbums: int = 5, 
 # ADD HOMEWORK - adds homework for a certain user
 @tree.command(name='add_homework', description='Add homework to someone\'s list', guild=my_guild)
 @app_commands.describe(entry="the artist - album you are trying to add (select an autocomplete option)",
-                       user='the user whose homework list you\'re adding to')
-@app_commands.autocomplete(entry=autocomplete_spotify)
+                       user="the user whose homework list you're adding to")
+@app_commands.autocomplete(entry=ac.autocomplete_spotify(spotify.search_album))
 async def add_homework(interaction: discord.Interaction, entry: str, user: discord.User = None):
     try:
         await interaction.response.defer()
@@ -846,8 +735,8 @@ async def add_homework(interaction: discord.Interaction, entry: str, user: disco
 
 
 # GET HOMEWORK - lists homework for a certain user
-@tree.command(name='get_homework', description='View someone\'s homework', guild=my_guild)
-@app_commands.describe(user='the user whose homework list you\'re looking at')
+@tree.command(name='get_homework', description="View someone's homework", guild=my_guild)
+@app_commands.describe(user="the user whose homework list you're looking at")
 async def get_homework(interaction: discord.Interaction, user: discord.User = None):
     try:
         await interaction.response.defer()
@@ -865,7 +754,7 @@ async def get_homework(interaction: discord.Interaction, user: discord.User = No
 # REMOVE_HOMEWORK: deletes homework from a users list
 @tree.command(name='remove_homework', description='remove homework from your list', guild=my_guild)
 @app_commands.describe(entry="the artist - album you are trying to remove from your list (see autocomplete)")
-@app_commands.autocomplete(entry=autocomplete_artist_album_homework_specific)
+@app_commands.autocomplete(entry=ac.autocomplete_artist_album_homework_specific(homework.get_homework, conn))
 async def remove_homework(interaction: discord.Interaction, entry: str):
     try:
         # fetches album from album_master and deletes it from the users homework table
@@ -880,7 +769,7 @@ async def remove_homework(interaction: discord.Interaction, entry: str):
 # ADD ALL HOMEWORK - adds homework to everyones homework
 @tree.command(name='add_all_homework', description='add homework to everyones list', guild=my_guild)
 @app_commands.describe(entry="the artist - album you are trying to add to everyones list (see autocomplete)")
-@app_commands.autocomplete(entry=autocomplete_spotify)
+@app_commands.autocomplete(entry=ac.autocomplete_spotify(spotify.search_album))
 async def add_all_homework(interaction: discord.Interaction, entry: str):
     try:
         await interaction.response.defer()
